Moved navbar out of shared
