+++ src/app/navbar/navbar.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { makeDecorator } from '@angular/core/src/util/decorators';
 
-import { NavItem } from '../services';
+import { NavItem } from './services';
 
 @Component({
   selector: 'app-navbar',
